Fix some issues in loan tests

+++ core/models.py
@@ -19,7 +19,7 @@ class Loan(models.Model):
     client = models.ForeignKey(Client, on_delete=models.PROTECT)
     amount = models.DecimalField(decimal_places=2, max_digits=10)
     term = models.IntegerField(validators=[validators.validate_term])
-    rate = models.DecimalField(decimal_places=2, max_digits=10)
+    rate = models.FloatField()
     date = models.DateTimeField(auto_now_add=True)
     installment = models.FloatField()
     paid = models.BooleanField(default=False)
